# src/generated/swagger.json
{
	"components": {
...
		"schemas": {
			"User": {
				"properties": {
					"isActive": {
						"type": "boolean"
					},
					"createAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"passwordHash": {
						"type": "string",
						"nullable": true
					},
					"email": {
						"type": "string",
						"nullable": true
					},
					"fullName": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"isActive",
					"createAt",
					"updatedAt",
					"passwordHash",
					"email",
					"fullName",
					"username",
					"id"
				],
				"type": "object",
				"description": "Model User"
			},
			"Photo": {
				"properties": {
					"isActive": {
						"type": "boolean"
					},
					"createAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"url": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"isActive",
					"createAt",
					"updatedAt",
					"userId",
					"url",
					"id"
				],
				"type": "object",
				"description": "Model Photo"
			},
			"Group": {
				"properties": {
					"isActive": {
						"type": "boolean"
					},
					"createAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"description": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"isActive",
					"createAt",
					"updatedAt",
					"description",
					"name",
					"id"
				],
				"type": "object",
				"description": "Model Group"
			},
			"UserDetailDto": {
				"allOf": [
					{
						"$ref": "#/components/schemas/User"
					},
					{
						"properties": {
							"permissions": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"groups": {
								"items": {
									"$ref": "#/components/schemas/Group"
								},
								"type": "array"
							},
							"photos": {
								"items": {
									"$ref": "#/components/schemas/Photo"
								},
								"type": "array"
							}
						},
						"required": [
							"permissions",
							"groups",
							"photos"
						],
						"type": "object"
					}
				]
			},
			"UserDto": {
				"$ref": "#/components/schemas/User"
			},
			"Permission": {
				"properties": {
					"isActive": {
						"type": "boolean"
					},
					"createAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"isActive",
					"createAt",
					"updatedAt",
					"description",
					"name",
					"id"
				],
				"type": "object",
				"description": "Model Permission"
			},
			"GroupDetailDto": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Group"
					},
					{
						"properties": {
							"permissions": {
								"items": {
									"$ref": "#/components/schemas/Permission"
								},
								"type": "array"
							}
						},
						"required": [
							"permissions"
						],
						"type": "object"
					}
				]
			},
			"UpdateGroupDto": {
				"properties": {
					"permissions": {
						"items": {
							"properties": {
								"name": {
									"type": "string"
								}
							},
							"required": [
								"name"
							],
							"type": "object"
						},
						"type": "array"
					},
					"isActive": {
						"type": "boolean"
					},
					"description": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"Paginate_Group_": {
				"properties": {
					"rows": {
						"items": {
							"$ref": "#/components/schemas/Group"
						},
						"type": "array"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"perPage": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"rows",
					"page",
					"perPage",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GroupListDto": {
				"$ref": "#/components/schemas/Paginate_Group_"
			},
			"CreateGroupDto": {
				"properties": {
					"permissions": {
						"items": {
							"properties": {
								"name": {
									"type": "string"
								}
							},
							"required": [
								"name"
							],
							"type": "object"
						},
						"type": "array"
					},
					"description": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"description",
					"name"
				],
				"type": "object"
			},
			"GroupDto": {
				"$ref": "#/components/schemas/Group"
			},
			"PermissionDetailDto": {
				"$ref": "#/components/schemas/Permission"
			},
			"PermissionDto": {
				"$ref": "#/components/schemas/Permission"
			},
			"UpdatePermissionDto": {
				"properties": {
					"isActive": {
						"type": "boolean"
					},
					"description": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"Paginate_Permission_": {
				"properties": {
					"rows": {
						"items": {
							"$ref": "#/components/schemas/Permission"
						},
						"type": "array"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"perPage": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"rows",
					"page",
					"perPage",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PermissionListDto": {
				"$ref": "#/components/schemas/Paginate_Permission_"
			},
			"CreatePermissionDto": {
				"properties": {
					"description": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object"
			},
...
			"UpdateUserDto": {
				"properties": {
					"photos": {
						"items": {
							"properties": {
								"url": {
									"type": "string"
								}
							},
							"required": [
								"url"
							],
							"type": "object"
						},
						"type": "array"
					},
					"groups": {
						"items": {
							"properties": {
								"name": {
									"type": "string"
								}
							},
							"required": [
								"name"
							],
							"type": "object"
						},
						"type": "array"
					},
					"isActive": {
						"type": "boolean"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"Paginate_User_": {
				"properties": {
					"rows": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"perPage": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"rows",
					"page",
					"perPage",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserListDto": {
				"$ref": "#/components/schemas/Paginate_User_"
			},
			"CreateUserDto": {
				"properties": {
					"email": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"username": {
						"type": "string",
						"minLength": 1
					}
				},
				"required": [
					"fullName",
					"username"
				],
				"type": "object"
			}
		},
		"securitySchemes": {
			"basic": {
				"scheme": "basic",
				"type": "http"
			},
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"name": "Bearer Authentication"
			}
		}
	},
	"info": {
		"title": "rest-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/me": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDetailDto"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"refreshToken": {
											"type": "string"
										},
										"accessToken": {
											"type": "string"
										}
									},
									"required": [
										"refreshToken",
										"accessToken"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"username": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"username"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/change_password": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"newpassword": {
										"type": "string"
									},
									"oldpassword": {
										"type": "string"
									}
								},
								"required": [
									"newpassword",
									"oldpassword"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/refresh_token": {
			"post": {
				"operationId": "PostRefreshToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"accessToken": {
											"type": "string"
										}
									},
									"required": [
										"accessToken"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"refreshToken": {
										"type": "string"
									}
								},
								"required": [
									"refreshToken"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/logout": {
			"get": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"401": {
						"description": "Success Logout and Unauthorized"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "refreshToken",
						"required": false,
						"schema": {
							"default": "None",
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/users/groups/{id}": {
			"get": {
				"operationId": "GetGroupDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GroupDetailDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"groups:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateGroup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GroupDetailDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"groups:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateGroupDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteGroup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GroupDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"groups:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/users/groups": {
			"get": {
				"operationId": "GetGroupList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GroupListDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "perPage",
						"required": false,
						"schema": {
							"default": 100,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "isActive",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateGroup",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GroupDetailDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"groups:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateGroupDto"
							}
						}
					}
				}
			}
		},
		"/api/v1/users/permissions/{id}": {
			"get": {
				"operationId": "GetPermissionDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PermissionDetailDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"permissions:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdatePermission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PermissionDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"permissions:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePermissionDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletePermission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PermissionDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"permissions:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/users/permissions": {
			"get": {
				"operationId": "GetPermissionList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PermissionListDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"permissions:read"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "perPage",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "username",
						"required": false,
						"schema": {
							"default": null,
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "isActive",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					}
				]
			},
			"post": {
				"operationId": "CreatePermission",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PermissionDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"permissions:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePermissionDto"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/users/{id}": {
			"get": {
				"operationId": "GetUserDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDetailDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDetailDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"userAdmin",
							"users:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"userAdmin",
							"users:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/users": {
			"get": {
				"operationId": "GetUserList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserListDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"userAdmin",
							"users:read"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "perPage",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "isActive",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						}
					}
				},
				"description": "Hola mundo",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"userAdmin",
							"users:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDto"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}